Reverted LB changes
